Added double for pulsepersec

+++ HU2/Can_Node_Module/Can_Node_Module_v6/Can_Node_Module_v6/ADC.h
@@ -5,6 +5,6 @@
 void int_ADC(void);
 void getADC(uint8_t ADC_pin);
 
-extern uint16_t PulsePerSec[4];
+extern double PulsePerSec[4];
 
 #endif
